feat: add async file target + fix logFileName
